# Implement plugins for logging drivers