--- a/api/types/plugins/logdriver/gen.go
+++ b/api/types/plugins/logdriver/gen.go
@@ -1,0 +1,3 @@
+//go:generate protoc --gogofast_out=import_path=github.com/docker/docker/api/types/plugins/logdriver:. entry.proto
+
+package logdriver