--- a/docs/extend/config.md
+++ b/docs/extend/config.md
@@ -59,6 +59,8 @@ Config provides the base accessible fields for working with V0 plugin format
 
         - **docker.authz/1.0**
 
+        - **docker.logdriver/1.0**
+
     - **`socket`** *string*
 
       socket is the name of the socket the engine should use to communicate with the plugins.