… api/types/plugins/logdriver/entry.pb.go
… vendor/github.com/gogo/protobuf/io/full.go
… vendor/github.com/gogo/protobuf/io/io.go
… vendor/github.com/gogo/protobuf/io/uint32.go
… vendor/github.com/gogo/protobuf/io/varint.go