This is intended as PoC code that as of yet untested other than ensuring compilation.

Logging plugins use the same HTTP interface as other plugins for basic
command operations meanwhile actual logging operations are handled (on
Unix) via a fifo. I believe this is much cleaner than dealing with go1.8 style plugins.

The plugin interface looks like so:

```go
type loggingPlugin interface {
  StartLogging(fifoPath string, loggingContext Context) error
  StopLogging(fifoPath)
}
```

This means a plugin must implement `LoggingDriver.StartLogging` and
`LoggingDriver.StopLogging` endpoints and be able to consume the passed
in fifo.

Logs are sent via stream encoder to the fifo encoded with json.
Ideally this would be protobuf, but I did not implement this yet.

ping @tiborvass @anusha-ragunathan