@cpuguy83 2016-11-15
The other alternative here would be to, instead of using fifo's, either hijack the HTTP stream (getting rid of the HTTP overhead), or potentially force an upgrade to HTTP2 (which is likely simpler for implementers than hijack).