@tianon 2016-11-17
Just to note for those looking for something in the interim, Docker _does_ currently support GELF (which is a standardized protocol for logging), so anything which can consume that (and/or translate that, like `logstash` input/output plugins) can be connected currently that way while a proper plugin interface is designed/developed. :+1: