@AkihiroSuda 2016-11-21
Thank you for working on this :+1: 

Some questions about FIFO, from my previous similar PR (https://github.com/docker/docker/pull/23198)

 * What happens if the plugin doesn't `read(2)` from the FIFO?  Probably the container gets stuck because its `write(2)` to stdout blocks?
 * What happens if the plugin crashes?