@cpuguy83 2016-11-22
> What happens if the plugin doesn't read(2) from the FIFO? Probably the container gets stuck because its write(2) to stdout blocks?

Correct, write will be blocked, which is the behavior for all existing containers except UDP based ones.

> What happens if the plugin crashes?

writes will be blocked until the plugin comes back