@cpuguy83 2016-12-14
Since we are already going to be communicating over a unix socket, I'm going to ditch the fifo. These offer similar performance, but with the unix socket I don't have to setup anything else.