@cpuguy83 2017-01-02
Just an update on this:

- writing logs works
- Kept usage of fifo since it is actually much simpler than dealing with the http connection, can supported named pipes on Windows later when there is platform support for this
- Reading logs is supported if the log driver reports that it supports reading

To think about (need input from others):
1. Should we just use a single stream for all log messages (from all containers) to the log driver? This would require encoding the log context with each message (easy enough). This would likely simplify plugin implementations, but make the docker side a bit more complex.
2. Do we need some cleanup call for log drivers? I'm thinking likely not but throwing it out there. As is I have a plugin which wraps the jsonfile driver which would need a way cleanup logs for removed containers... but I think this is just because of the nature of the jsonfile logger, a real log driver wouldn't just remove logs in this manner.

Here's my plugin implementation: https://github.com/cpuguy83/docker-log-driver-test
I do have a (simple) integration test that uses this driver to both write and read back container logs.