@LK4D4 2017-01-26
I'm ok with design
ping @docker/core-maintainers for moving to code-review.