@cpuguy83 2017-03-05
Ok, this is updated to use protobuf encoding.
The framing for the stream is from `gogo/protobuf/io` which is similar to grpc, where it sends `[buf size][buf]`.
I've also updated my demo log plugin to use this format.

Please check the `.proto`. I guess we should probably move this into `/api` somewhere?