@thaJeztah 2017-03-05
Looks like golint doesn't like it

```
18:00:33 Errors from golint:
18:00:33 daemon/logger/plugin.pb.go:5:1: package comment should not have leading space
18:00:33 daemon/logger/plugin.pb.go:19:8: a blank import should be only in a main or test package, or have a comment justifying it