@AkihiroSuda 2017-03-06
Just a question, is there performance comparison between json and protobuf(+gogo)?