@cpuguy83 2017-03-06
@AkihiroSuda I can certainly do it.
The json encoder is fairly alloc heavy and CPU intensive.

We have a bunch of hacks for `jsonfile` to get around this which I would prefer to not propagate here. I can test `github.com/ugorji/codec`'s json encoder as well as `encoding/json`