@cpuguy83 2017-03-08
```
BenchmarkMessageProtobufEncoder-8        	20000000	      2697 ns/op	     128 B/op	       6 allocs/op
BenchmarkMessageStdlibJSONEncoder-8      	10000000	      4030 ns/op	     232 B/op	       5 allocs/op
BenchmarkMessageUgorjiJSONEncoder-8      	 1000000	     43477 ns/op	     177 B/op	      33 allocs/op
BenchmarkMessageUgorjiMsgPackEncoder-8   	 2000000	     24874 ns/op	     143 B/op	      19 allocs/op
```

https://gist.github.com/cpuguy83/a2e8e22e1a70db5476abb729999ace7a