@cpuguy83 2017-03-08
Seriously?
```
18:50:08 The result of vndr differs
18:50:08 
18:50:08  M vendor/github.com/seccomp/libseccomp-golang/README
```