@cpuguy83 2017-03-22
Fixed code comments.