@cpuguy83 2017-03-29
wrt the benchmarks listed above, I found some issues with the gogo encoder that should reduce allocs by at least 3.