@cpuguy83 2017-04-05
- Some changes to how encoder/decoder works (-3 allocs)
- Change from time.Time to int64 (time.UnixNano()) (-3 allocs) --- also changed field name to `TimeNano`

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkMessageProtobufEncoder-8     2697          990           -63.29%

benchmark                             old allocs     new allocs     delta
BenchmarkMessageProtobufEncoder-8     6              0              -100.00%

benchmark                             old bytes     new bytes     delta
BenchmarkMessageProtobufEncoder-8     128           0             -100.00%
```