@thaJeztah 2017-04-05
```
18:55:23 ----------------------------------------------------------------------
18:55:23 FAIL: docker_cli_plugins_logdriver_test.go:10: DockerSuite.TestPluginLogDriver
18:55:23 
18:55:23 docker_cli_plugins_logdriver_test.go:18:
18:55:23     c.Assert(strings.TrimSpace(out), checker.Equals, "hello")
18:55:23 ... obtained string = ""
18:55:23 ... expected string = "hello"
18:55:23 
18:55:27 
18:55:27 ----------------------------------------------------------------------
```