@cpuguy83 2017-04-06
"These are not the bugs you are looking for"

Issue was that my plugin needed to be updated to account for the changed field.