@ehazlett 2017-04-07
Do we need to update `github.com/gogo/protobuf` in vendor.conf?