@ehazlett 2017-04-07
Should we add tooling to the makefile to generate the protos like we have for containerd / swarmkit (https://github.com/containerd/containerd/blob/master/Makefile#L45)?  We are most likely going to use protos for the generic runtime support as well.  This might be better in a follow up.