@anusha-ragunathan 2017-04-10
`docs/extend/config.md` has the list of drivers currently supported. It needs to be updated.