@anusha-ragunathan 2017-04-10
@cpuguy83: I'm okay with a follow up PR to address 
- `api/types` reorg
- swarm services plumbing for scheduling where logging plugins are available.  Thenwe can also add swarm tests in https://github.com/docker/docker/blob/master/integration-cli/docker_cli_swarm_unix_test.go