@thaJeztah 2017-04-10
Discussed with @cpuguy83 and we'll do a follow up for docs due to code-freeze today